Include profile description
Utilizing more information available in the profile, a sensible
description is readily available.

According to the API description [1] an "email" field should also be
available. However, I could not find this field filled in a sample of
randomly selected Vine users, so I assume it is only visible when logged
in or when examining the own account. It thus an be regarded unfilled
for the purposes of vinefeed.

[1]: https://github.com/starlock/vino/wiki/API-Reference#get-user-data